Renamed the MarginTrading.AccountsManagement.DomainModels namespace to MarginTrading.AccountsManagement.InternalModels
