+++ src/MarginTrading.AccountsManagement/Controllers/AccountsController.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using MarginTrading.AccountsManagement.Contracts;
 using MarginTrading.AccountsManagement.Contracts.Api;
 using MarginTrading.AccountsManagement.Contracts.Models;
-using MarginTrading.AccountsManagement.DomainModels;
+using MarginTrading.AccountsManagement.InternalModels;
 using MarginTrading.AccountsManagement.Infrastructure;
 using MarginTrading.AccountsManagement.Services;
 using Microsoft.AspNetCore.Mvc;
