+++ src/MarginTrading.AccountsManagement/InternalModels/AccountHistoryType.cs
@@ -1,4 +1,4 @@
-﻿namespace MarginTrading.AccountsManagement.DomainModels
+﻿namespace MarginTrading.AccountsManagement.InternalModels
 {
     public enum AccountHistoryType
     {
